feat: declutter App.tsx
